ScrapperWebApp/requirements.txt
# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
supabase>=2.0.0
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0

# Web scraping dependencies (from Scraping folder)
selenium>=4.15.0
undetected-chromedriver>=3.5.4
beautifulsoup4>=4.12.2
requests>=2.31.0
webdriver-manager>=4.0.1
fake-useragent>=1.4.0
lxml>=4.9.3

# Python 3.12+ compatibility
setuptools>=68.0.0
urllib3>=2.0.0
certifi>=2023.7.22
